Updates Makefile to be more useful
The original makefile used a non-standard FORTRAN77 compiler, did not
follow "the usual conventions" and had a less-than-useful 'dist' rule.

This commit updates the Makefile to use the GNU gfortran compiler, which
is generally recommended for its strictness and flexibility.  It also
adopts traditional conventions for variable usage (OBJS, EXE, OBJS,
etc.) and implements a more useful 'dist' rule.  It also adds a 'clean'
rule.