Adds install rules, add more standard variables
In doing a little reading, I remembered that the "dist" rule is for
creating a distribution file and there is an "install" rule to put the
file(s) into the appropriate installation directory and verify file
protections.  Also, that FC is just for the name of the compiler and
FFLAGS is for the flags.  The implied compile rules use this convention.
(See GNU Make Manual, Automatic Variables) for details.

This commit adds the standard variables, adds a tar command to create
the distribution file and adds a stub for the install rule.
ZZ